refactor: move panes to dedicated lib
