+++ extensions/pyRevitTools.extension/lib/panes/clipboard/clipboard.Icons.xaml
