+++ extensions/pyRevitTools.extension/lib/panes/clipboard/clipboard_content.xaml
