+++ extensions/pyRevitTools.extension/lib/panes/clipboard/clipboard_page.xaml
