+++ extensions/pyRevitTools.extension/lib/panes/clipboard/clipboard_ui.ResourceDictionary.de_de.xaml
