+++ extensions/pyRevitTools.extension/lib/panes/clipboard/clipboard_ui.ResourceDictionary.en_us.xaml
