+++ extensions/pyRevitTools.extension/lib/panes/clipboard/clipboard_ui.ResourceDictionary.es_es.xaml
