+++ extensions/pyRevitTools.extension/lib/panes/clipboard/clipboard_ui.ResourceDictionary.pt_br.xaml
