+++ extensions/pyRevitTools.extension/lib/panes/clipboard/clipboard_ui.ResourceDictionary.ru.xaml
