+++ extensions/pyRevitTools.extension/lib/panes/clipboard/clipboard_window.xaml
