+++ extensions/pyRevitTools.extension/lib/panes/clipboard/pane.py
