+++ extensions/pyRevitTools.extension/lib/panes/customprops/pane_ui.ResourceDictionary.chinese_s.xaml
