+++ extensions/pyRevitTools.extension/lib/panes/customprops/pane_ui.ResourceDictionary.en_us.xaml
