+++ extensions/pyRevitTools.extension/lib/panes/customprops/pane_ui.ResourceDictionary.fr_fr.xaml
