+++ extensions/pyRevitTools.extension/lib/panes/customprops/pane_ui.ResourceDictionary.pt_br.xaml
