+++ extensions/pyRevitTools.extension/pyRevit.tab/Modify.panel/edit1.stack/Match.splitpushbutton/Match History Clipboard.smartbutton/script.py
@@ -3,7 +3,7 @@ from pyrevit.revit import ui
 from pyrevit.framework import Threading, System
 import pyrevit.extensions as exts
 from pyrevit.coreutils.ribbon import ICON_MEDIUM
-from match import clipboard
+from panes.clipboard.pane import MatchHistoryClipboard
 
 
 def __selfinit__(script_cmp, ui_button_cmp, __rvt__):
@@ -19,8 +19,8 @@ def __selfinit__(script_cmp, ui_button_cmp, __rvt__):
 
     def update_icon_from_pane():
         try:
-            if forms.is_registered_dockable_panel(clipboard.MatchHistoryClipboard):
-                dockable = forms.get_dockable_panel(clipboard.MatchHistoryClipboard)
+            if forms.is_registered_dockable_panel(MatchHistoryClipboard):
+                dockable = forms.get_dockable_panel(MatchHistoryClipboard)
                 set_icon(dockable.IsShown())
             else:
                 set_icon(False)
@@ -55,8 +55,8 @@ def __selfinit__(script_cmp, ui_button_cmp, __rvt__):
 
 
 if __name__ == "__main__":
-    if forms.is_registered_dockable_panel(clipboard.MatchHistoryClipboard):
-        dockable_panel = forms.get_dockable_panel(clipboard.MatchHistoryClipboard)
+    if forms.is_registered_dockable_panel(MatchHistoryClipboard):
+        dockable_panel = forms.get_dockable_panel(MatchHistoryClipboard)
         forms.toggle_dockable_panel(
-            clipboard.MatchHistoryClipboard, not dockable_panel.IsShown()
+            MatchHistoryClipboard, not dockable_panel.IsShown()
         )
