+++ extensions/pyRevitTools.extension/startup.py
@@ -2,17 +2,18 @@ from pyrevit._perf import mark as _perfmark, time_block as _perfblock
 _perfmark("startup.pyRevitTools:entry")
 
 from pyrevit import forms, script
-from customprops.custom_props_pane import CONFIG_SECTION, CustomPropertiesPanel
 _perfmark("startup.pyRevitTools:after `from pyrevit import forms, script`")
+from panes.customprops.pane import CONFIG_SECTION, CustomPropertiesPanel
+_perfmark("startup.pyRevitTools:after `from panes.customprops.pane import CONFIG_SECTION, CustomPropertiesPanel`")
 
-from match import clipboard
-_perfmark("startup.pyRevitTools:after `from match import panel`")
+from panes.clipboard.pane import MatchHistoryClipboard
+_perfmark("startup.pyRevitTools:after `from match import MatchHistoryClipboard`")
 
 logger = script.get_logger()
 
-if not forms.is_registered_dockable_panel(clipboard.MatchHistoryClipboard):
+if not forms.is_registered_dockable_panel(MatchHistoryClipboard):
     with _perfblock("startup.pyRevitTools:register_dockable_panel(MatchHistoryClipboard)"):
-        forms.register_dockable_panel(clipboard.MatchHistoryClipboard, default_visible=False)
+        forms.register_dockable_panel(MatchHistoryClipboard, default_visible=False)
 else:
     logger.debug("Skipped registering dockable pane. Already exists.")
